requirements.txt
streamlit==1.9.0

absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cshogi==0.3.6
cycler==0.11.0
decorator==5.1.1
executing==0.8.3
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.44.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
ipython==8.2.0
jedi==0.18.1
kiwisolver==1.4.2
Markdown==3.3.6
matplotlib==3.5.2
matplotlib-inline==0.1.3
multidict==6.0.2
numpy==1.22.3
oauthlib==3.2.0
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytorch-lightning==1.6.3
pytz==2022.1
PyYAML==6.0
pyvips==2.2.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
svgwrite==1.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.10.1+cpu
torchvision==0.11.2+cpu
setuptools==59.5.0
torchmetrics==0.7.3
tqdm==4.63.1
traitlets==5.1.1
typing-extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.0.3
whitenoise==6.1.0
yarl==1.7.2
zipp==3.7.0